fix(app): allow fractional charging wattage on smartphones
Pydantic rejects records like Apple iPhone XS Max (7.5 W wireless) and
LG G2 (4.5 W wireless) because the schema/model declared
`charging_wired_w` and `charging_wireless_w` as `int`. Industry specs
routinely use fractional watts (5 / 7.5 / 10 / 15 W Qi, 4.5 W early
LG), so widen both fields to `float | None`.

Fixes the `int_from_float` failure that took down test, deploy-pages,
and refresh-data on f510014.

Refs #1

+++ app/models/smartphone.py
@@ -39,8 +39,8 @@ class Smartphone(SQLModel, table=True):
 
     # Battery
     battery_mah: int
-    charging_wired_w: int | None = None
-    charging_wireless_w: int | None = None
+    charging_wired_w: float | None = None
+    charging_wireless_w: float | None = None
 
     # Physical
     weight_g: float
